move vault example into own mod; split away program definition into separate file
